Test fails on jumping SIMD_WIDTH slots in find

+++ include/flatpack/entry64.hpp
@@ -5,13 +5,13 @@
 
 namespace flatpack {
 
-struct Entry64 {          // Padded to 24 bytes anyway by Compiler
-  uint64_t key;           // 8 bytes
-  uint64_t value;         // 8 bytes
-  uint8_t fingerprint;    // 1 byte => Fast compare
-  uint8_t probe_distance; // 1 byte => Robin Hood algo
-  uint8_t reserved;       // 1 byte
-                          // = 19 bytes > 16 bytes
+struct Entry64 {           // Padded to 24 bytes anyway by Compiler
+  uint64_t key;            // 8 bytes
+  uint64_t value;          // 8 bytes
+  uint8_t fingerprint;     // 1 byte => Fast compare
+  uint16_t probe_distance; // 2 byte => Robin Hood algo
+  uint8_t reserved;        // 1 byte
+                           // = 20 bytes > 16 bytes < 24 bytes
 };
 
 } // namespace flatpack
